refactor: split upgrade service.ts into focused modules
Break down the 775-line monolithic upgrade/service.ts into 9 modules:
- types.ts, constants.ts, github.ts, checker.ts, checksum.ts,
  download.ts, apply.ts, files.ts, and a slim orchestration service.ts

Also fixes review findings:
- Deduplicate filename regex (route now imports VALID_FILE_NAME_RE)
- Expand download guard to cover 'verifying' state
- Log download errors instead of silently discarding
- Harden tar extraction with --no-same-owner --no-overwrite-dir
- Add tests for checksum and file management modules
